refactor: clean up db_resilience module and update query type naming
